+++ Examples/ISHDeploy/Set-ADFSIntegration.ImplicitRemoting.ps1
@@ -1,6 +1,8 @@
 param (
     [Parameter(Mandatory=$false)]
     [string[]]$Computer,
+    [Parameter(Mandatory=$false)]
+    [pscredential]$Credential=$null,
     [Parameter(Mandatory=$true)]
     [string]$DeploymentName,
     [Parameter(Mandatory=$true)]
@@ -37,7 +39,7 @@ try
     if($Computer)
     {
         $ishDelpoyModuleName="ISHDeploy.$ISHVersion"
-        $remote=Add-ModuleFromRemote -ComputerName $Computer -Name $ishDelpoyModuleName
+        $remote=Add-ModuleFromRemote -ComputerName $Computer -Credential $Credential -Name $ishDelpoyModuleName
     }
     $remoteADFS=Add-ModuleFromRemote -ComputerName $adfsComputerName -Name ADFS